resume-screening/requirements.txt
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
sentence-transformers==2.7.0
pypdf==4.2.0
python-docx==1.1.2